requirements.txt
absl-py==0.13.0
addict==2.4.0
aiofiles==0.8.0
antlr4-python3-runtime==4.8
anyio==3.3.0
appdirs==1.4.4
argcomplete==1.12.3
argon2-cffi==20.1.0
async-generator==1.10
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
backports.zoneinfo==0.2.1
bleach==4.0.0
boto3==1.20.23
botocore==1.23.23
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.9
cloudpickle==1.6.0
cycler==0.10.0
Cython==0.29.23
debugpy==1.4.1
decorator==5.0.9
defusedxml==0.7.1
Deprecated==1.2.13
deprecation==2.1.0
detectron2==0.4+cu111
dill==0.3.4
dnspython==2.1.0
entrypoints==0.3
eventlet==0.33.0
fiftyone==0.14.2
fiftyone-brain==0.7.2
fiftyone-db==0.3.0
future==0.18.2
fvcore==0.1.3.post20210317
glob2==0.7
google-auth==1.32.1
google-auth-oauthlib==0.4.4
greenlet==1.1.2
grpcio==1.38.1
h11==0.12.0
httpcore==0.14.3
httpx==0.21.1
idna==2.10
imageio==2.9.0
iopath==0.1.9
ipykernel==6.0.3
ipython==7.26.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==3.0.1
jmespath==0.10.0
joblib==1.0.1
json5==0.9.6
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyter-packaging==0.10.4
jupyter-server==1.10.2
jupyterlab==3.1.1
jupyterlab-pygments==0.1.2
jupyterlab-server==2.6.2
jupyterlab-widgets==1.0.0
kaleido==0.2.1
kiwisolver==1.3.1
llvmlite==0.37.0
Mako==1.1.5
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mistune==0.8.4
mongoengine==0.20.0
motor==2.5.1
nbclassic==0.3.1
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
ndjson==0.3.1
nest-asyncio==1.5.1
networkx==2.6.3
notebook==6.4.0
ntplib==0.4.0
numba==0.54.1
numpy==1.20.3
oauthlib==3.1.1
omegaconf==2.1.0
open3d==0.13.0
opencv-contrib-python==4.5.4.60
opencv-python-headless==4.5.4.60
packaging==21.0
pandas==1.3.1
pandocfilters==1.4.3
parso==0.8.2
patool==1.12
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.1
plotly==4.14.3
portalocker==2.3.0
pprintpp==0.4.0
probreg==0.3.4
prometheus-client==0.11.0
prompt-toolkit==3.0.19
protobuf==3.17.3
psutil==5.8.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.8.0
pycocotools==2.0.2
pycparser==2.20
pycuda==2021.1
pydot==1.4.2
Pygments==2.9.0
pymongo==3.12.3
pyparsing==2.4.7
pyrealsense2==2.48.0.3381
pyrsistent==0.18.0
python-dateutil==2.8.1
pytools==2021.2.8
pytz==2021.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==22.1.0
qtm==2.1.1
requests==2.25.1
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
retrying==1.3.3
rfc3986==1.5.0
rsa==4.7.2
s3transfer==0.5.0
scikit-image==0.18.3
scikit-learn==0.24.2
scipy==1.7.1
Send2Trash==1.7.1
Shapely==1.8.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
tabulate==0.8.9
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
termcolor==1.1.0
terminado==0.10.1
testpath==0.5.0
threadpoolctl==2.2.0
tifffile==2021.10.12
tomlkit==0.7.2
torch==1.8.1+cu111
torchaudio==0.8.1
torchvision==0.9.1+cu111
tornado==6.1
tqdm==4.61.2
traitlets==5.0.5
transforms3d==0.3.1
typing-extensions==3.10.0.0
tzdata==2021.5
tzlocal==4.1
universal-analytics-python3==1.1.1
urllib3==1.26.6
voxel51-eta==0.6.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.1.0
Werkzeug==2.0.1
widgetsnbextension==3.5.1
wrapt==1.13.3
xmltodict==0.12.0
yacs==0.1.8
